Renamed signing_workflow to sign_workflow.
